requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
cryptography==46.0.2
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.6.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.15.2
docker==7.1.0
idna==3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.10
pycparser==2.23
pyOpenSSL==25.3.0
python-dotenv==1.0.1
pywin32==311
requests==2.32.3
ruff==0.13.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.16
Werkzeug==3.1.3
